+++ ft_printf/Libft/source/ft_putstr_fd.c
@@ -6,15 +6,15 @@
 /*   By: eunwolee <eunwolee@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/15 16:25:54 by eunwolee          #+#    #+#             */
-/*   Updated: 2022/12/17 17:03:39 by eunwolee         ###   ########.fr       */
+/*   Updated: 2022/12/19 15:48:22 by eunwolee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-ssize_t	ft_putstr_fd(char *s, int fd)
+int	ft_putstr_fd(char *s, int fd)
 {
 	if (fd < 0)
 		return (-1);
-	return (write(fd, s, ft_strlen(s)));
+	return ((int)write(fd, s, ft_strlen(s)));
 }
