+++ ft_printf/source/get_string.c
@@ -6,17 +6,17 @@
 /*   By: eunwolee <eunwolee@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/17 15:02:11 by eunwolee          #+#    #+#             */
-/*   Updated: 2022/12/18 15:16:00 by eunwolee         ###   ########.fr       */
+/*   Updated: 2022/12/19 15:50:10 by eunwolee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/ft_printf.h"
 
-ssize_t	get_string(char fmt, va_list *ap)
+int	get_string(char fmt, va_list *ap)
 {
 	char	*str;
 	char	*(*func)(va_list *);
-	ssize_t	size;
+	int		size;
 
 	if (!ft_strchr("cspdiuxX", fmt))
 		return (ft_putchar_fd(fmt, 1));
